# perf: getting `ape console` to launch faster