--- a/src/ape/cli/options.py
+++ b/src/ape/cli/options.py
@@ -528,6 +528,10 @@ def incompatible_with(incompatible_opts) -> type[click.Option]:
     return IncompatibleOption
 
 
+def _project_path_callback(ctx, param, val):
+    return Path(val) if val else Path.cwd()
+
+
 def _project_callback(ctx, param, val):
     if "--help" in sys.argv or "-h" in sys.argv:
         # Perf: project option is eager; have to check sys.argv to
@@ -560,10 +564,12 @@ def _project_callback(ctx, param, val):
 
 
 def project_option(**kwargs):
+    _type = kwargs.pop("type", None)
+    callback = _project_path_callback if issubclass(_type, Path) else _project_callback
     return click.option(
         "--project",
         help="The path to a local project or manifest",
-        callback=_project_callback,
+        callback=callback,
         metavar="PATH",
         is_eager=True,
         **kwargs,