… src/ape/managers/project.py
… src/ape_console/_cli.py
… src/ape_test/provider.py
… tests/functional/test_cli.py
… tests/functional/test_console.py
… tests/functional/test_ecosystem.py