### What I did

doing what i can to get `ape console` to launch faster.
Mostly centers around connecting to networks faster
Also some other small perfs...
Am still working on the main bottle neck..

Before:

```
(ape_no_plugins) ➜  ape git:(main) time ape console

In [1]: exit
ape console  4.79s user 0.61s system 91% cpu 5.916 total
(ape_no_plugins) ➜  ape git:(main) time ape console

In [1]: exit
ape console  4.49s user 0.55s system 94% cpu 5.333 total
```

After:

```
ape console  3.87s user 0.64s system 101% cpu 4.460 total
(ape) ➜  ape git:(perf/console) time ape console

In [1]: exit
ape console  3.74s user 0.57s system 100% cpu 4.274 total
```

obviously this network dependent and such

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated